% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mvn.R
\name{tidy_csmvn}
\alias{tidy_csmvn}
\title{Cross-sectional Gaussian probabilistic reconciliation}
\usage{
tidy_csmvn(models, ...)
}
\arguments{
\item{models}{A column of models in a mable.}

\item{...}{Additional arguments passed to \code{\link[FoReco:csmvn]{FoReco::csmvn()}}, except for
\code{base}, \code{agg_mat}, \code{cons_mat}, and \code{res}, which are supplied internally.}
}
\value{
An object of class \code{lst_csmvn_mdl}
}
\description{
Reconciles a hierarchy or grouped time series using the \code{\link[FoReco:csmvn]{FoReco::csmvn()}}
function. The response variable of the hierarchy must be aggregated using
sums. Forecasted time points must match for all series in the hierarchy.
}
