% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_opt.R
\name{tidy_csrec}
\alias{tidy_csrec}
\title{Optimal combination cross-sectional reconciliation}
\usage{
tidy_csrec(models, ...)
}
\arguments{
\item{models}{A column of models in a mable.}

\item{...}{Additional arguments passed to \code{\link[FoReco:csrec]{FoReco::csrec()}}, except for
\code{base}, \code{agg_mat}, \code{cons_mat}, and \code{res}, which are supplied internally.}
}
\value{
An object of class \code{lst_csrec_mdl}
}
\description{
Reconciles a hierarchy or grouped time series using the \code{\link[FoReco:csrec]{FoReco::csrec()}}
function. The response variable of the hierarchy must be aggregated using
sums. Forecasted time points must match for all series in the hierarchy.
}
