% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_smp.R
\name{tidy_cssmp}
\alias{tidy_cssmp}
\title{Cross-sectional probabilistic reconciliation (sample approach)}
\usage{
tidy_cssmp(models, simulate = TRUE, bootstrap = FALSE, times = 1000, ...)
}
\arguments{
\item{models}{A column of models in a mable.}

\item{simulate}{Should forecasts be based on simulated future paths instead
of analytical results.}

\item{bootstrap}{Should innovations from simulated forecasts be bootstrapped
from the model's fitted residuals. This allows the forecast distribution to
have a different underlying shape which could better represent the nature
of your data.}

\item{times}{The number of future paths for simulations if \code{simulate = TRUE}.}

\item{...}{Additional arguments passed to \code{\link[FoReco:cssmp]{FoReco::cssmp()}}, except for
\code{sample}, \code{agg_mat}, \code{cons_mat}, and \code{res}, which are supplied internally.}
}
\value{
An object of class \code{lst_cssmp_mdl}
}
\description{
Reconciles a hierarchy or grouped time series using the \code{\link[FoReco:cssmp]{FoReco::cssmp()}}
function. The response variable of the hierarchy must be aggregated using
sums. Forecasted time points must match for all series in the hierarchy.
}
